requirements.txt
Flask>=2.0.0
Flask-SQLAlchemy>=2.5.0
python-dotenv>=0.19.0
pandas>=1.3.0
openpyxl>=3.0.0
langchain>=0.0.200
openai>=0.27.0
six>=1.16.0
python-dateutil>=2.8.2
langchain_community>=0.0.200